removed the last unnessecary static_casts

+++ src/Model/Power/Vphinode.cpp
@@ -12,7 +12,7 @@ namespace Model::Networkproblem::Power {
                           Eigen::VectorXd const & // last_state
                           ,
                           Eigen::VectorXd const &new_state) {
-    int index1 = static_cast<int>(get_start_state_index());
+    int index1 = get_start_state_index();
     int index2 = index1 + 1;
     rootfunction[index1] = new_state[index1] - boundaryvalue(new_time)[0];
 
@@ -27,7 +27,7 @@ namespace Model::Networkproblem::Power {
                                            Eigen::VectorXd const &,
                                            Eigen::VectorXd const & // new_state
   ) {
-    int index1 = static_cast<int>(get_start_state_index());
+    int index1 = get_start_state_index();
     int index2 = index1 + 1;
     jacobianhandler->set_coefficient(index1, index1, 1.0);
     jacobianhandler->set_coefficient(index2, index2, 1.0);
